requirements.txt
# Core LangChain packages
langchain
langchain-openai
langchain-community
langchain_experimental
langchain-huggingface

# OpenAI and Azure OpenAI
openai

# Streamlit for web interface
streamlit

# Environment variables
python-dotenv

# Document processing
unstructured
docx2txt
openpyxl
PyPDF2

# Vector database
faiss-cpu

# Data processing
pandas
numpy

# NLP utilities
nltk
sentence_transformers

# Database (if needed for other features)
sqlalchemy
psycopg2-binary

# Cloud storage (if needed)
dropbox

# Utilities
tqdm
pickle-mixin

# Development tools
pylint
black

# Additional dependencies that might be needed
tiktoken
pydantic
requests
urllib3